INTYGFV-12569: Moved encoding filter. Created tests for filter.
