+++ web/src/main/resources/webcert-config.xml
@@ -168,7 +168,7 @@
         <property name="savedRequestFactory" ref="savedRequestFactory"/>
     </bean>
 
-    <bean id="defaultCharacterEncodingFilter" class="se.inera.intyg.webcert.web.auth.DefaultCharacterEncodingFilter" />
+    <bean id="defaultCharacterEncodingFilter" class="se.inera.intyg.webcert.web.web.filter.DefaultCharacterEncodingFilter" />
 
     <!-- Eleg login redirect filter for easier URL handling in PP -->
     <bean id="elegRedirectFilter" class="se.inera.intyg.webcert.web.auth.eleg.ElegRedirectFilter" />
